Refactor: refactor login page container, presenter, types (interface) 파일 분리
