+++ coin/pages/login/index.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "../../src/components/commons/login/Login";
+import LoginPage from "../../src/components/commons/login/main/Login.container";
 export default function Login() {
     return (
         <LoginPage />
