+++ coin/src/components/commons/login/main/Login.container.tsx
@@ -1,0 +1,8 @@
+import LoginPageUI from "./Login.presenter";
+
+export default function LoginPage(): JSX.Element {
+    
+    return (
+        <LoginPageUI/>
+    );
+}
